+++ notebooks/noj_book/smile_others.clj
@@ -1,3 +1,9 @@
+;; # Smile other models reference - DRAFT 🛠
+
+;; Note that this chapter reqiures `scicloj.ml.smile` as an additional
+;; dependency to Noj.
+;; [![Clojars Project](https://img.shields.io/clojars/v/org.scicloj/scicloj.ml.smile.svg)](https://clojars.org/org.scicloj/scicloj.ml.smile)
+
 ^:kindly/hide-code
 (ns noj-book.smile-others
   (:require
@@ -6,7 +12,6 @@
    [scicloj.ml.smile.projections]
    [noj-book.utils.render-tools :refer [render-key-info]]))
 
-;; ## Smile other models reference
 ;; In the following we have a list of all model keys of Smile model-like
 ;; algorithms including parameters.
 ;; They can be used in the same way as other models:
@@ -19,17 +24,16 @@
 ;; Some do not support `ml/predict` and are defined as `unsupervised` learners.
 ;; Clustering and PCA are in this group.
 
-
 ;; ## Smile manifolds
 
 ^:kindly/hide-code
 (render-key-info :smile.manifold)
 
-;; ### Smile/Fastmath clustering
+;; ## Smile/Fastmath clustering
 ^:kindly/hide-code
 (render-key-info :fastmath.cluster)
 
-;; ### Smile projections
+;; ## Smile projections
 ^:kindly/hide-code
 (render-key-info :smile.projections)