+++ notebooks/noj_book/transformer_references.clj
@@ -1,3 +1,10 @@
+;; # Transformer reference  - DRAFT 🛠
+
+;; Note that this chapter reqiures `scicloj.ml.smile` as an additional
+;; dependency to Noj.
+;; [![Clojars Project](https://img.shields.io/clojars/v/org.scicloj/scicloj.ml.smile.svg)](https://clojars.org/org.scicloj/scicloj.ml.smile)
+
+
 (ns noj-book.transformer-references
   (:require
    [scicloj.kindly.v4.api :as kindly]
@@ -30,9 +37,6 @@
        (kind/md "__Clojure doc__:\n")
        (kind/md (:doc m))
        (kind/md "----------------------------------------------------------")]))))
-
-  
-;; ## Transformer reference
 
 (docu-fn (var nlp/count-vectorize))